ABCP Determination corrected
This commit corresponds to issue 'ABCP Pass determination is not correct. #1 '

Made item 'Hips' stick on itemList.
All 5 items are static, but just ItemRecyclerAdapter shows selectively.

ItemRecyclerAdapter.itemCount has been implemented, and it is changed on ABCPFragment when it needs to be. It might be decided by return value of ABCPRecord.invalidate (ArrayList<Item> items)

Still bodyfat requires 3rd state 'N/A', and needs to be reflected on view.